+++ edc-tests/e2e-tests/src/test/java/org/eclipse/tractusx/edc/lifecycle/MultiRuntimeTest.java
@@ -19,44 +19,47 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 
 import java.util.HashMap;
 
-import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.IDS_PATH;
+import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.DSP_PATH;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_CONNECTOR_PATH;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_CONNECTOR_PORT;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_DATAPLANE_CONTROL_PORT;
-import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_IDS_API;
-import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_IDS_API_PORT;
+import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_DSP_API_PORT;
+import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_DSP_CALLBACK;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_MANAGEMENT_PATH;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_MANAGEMENT_PORT;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.PLATO_PUBLIC_API_PORT;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_CONNECTOR_PATH;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_CONNECTOR_PORT;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_DATAPLANE_CONTROL_PORT;
-import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_IDS_API;
-import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_IDS_API_PORT;
+import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_DSP_API_PORT;
+import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_DSP_CALLBACK;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_MANAGEMENT_PATH;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_MANAGEMENT_PORT;
 import static org.eclipse.tractusx.edc.lifecycle.TestRuntimeConfiguration.SOKRATES_PUBLIC_API_PORT;
 
 
 public class MultiRuntimeTest {
-
+    public static final String BPN_SUFFIX = "-BPN";
+    public static final String SOKRATES_NAME = "SOKRATES";
+    public static final String SOKRATES_BPN = SOKRATES_NAME + BPN_SUFFIX;
 
     @RegisterExtension
     protected static Participant sokrates = new Participant(
             ":edc-tests:runtime",
-            "SOKRATES",
+            SOKRATES_NAME,
+            SOKRATES_BPN,
             new HashMap<>() {
                 {
                     put("edc.connector.name", "sokrates");
-                    put("edc.ids.id", "urn:connector:sokrates");
+                    put("edc.participant.id", SOKRATES_BPN);
                     put("web.http.port", String.valueOf(SOKRATES_CONNECTOR_PORT));
                     put("web.http.path", SOKRATES_CONNECTOR_PATH);
                     put("web.http.management.port", String.valueOf(SOKRATES_MANAGEMENT_PORT));
                     put("web.http.management.path", SOKRATES_MANAGEMENT_PATH);
-                    put("web.http.ids.port", String.valueOf(SOKRATES_IDS_API_PORT));
-                    put("web.http.ids.path", IDS_PATH);
+                    put("web.http.protocol.port", String.valueOf(SOKRATES_DSP_API_PORT));
+                    put("web.http.protocol.path", DSP_PATH);
+                    put("edc.dsp.callback.address", SOKRATES_DSP_CALLBACK);
                     put("edc.api.auth.key", "testkey");
-                    put("ids.webhook.address", SOKRATES_IDS_API);
                     put("web.http.public.path", "/api/public");
                     put("web.http.public.port", SOKRATES_PUBLIC_API_PORT);
 
@@ -70,25 +73,29 @@ public class MultiRuntimeTest {
                     put("edc.dataplane.selector.httpplane.properties", "{\"publicApiUrl\":\"http://localhost:" + SOKRATES_PUBLIC_API_PORT + "/api/public\"}");
                     put("edc.receiver.http.dynamic.endpoint", "http://localhost:" + SOKRATES_CONNECTOR_PORT + "/api/consumer/datareference");
                     put("tractusx.businesspartnervalidation.log.agreement.validation", "true");
+                    put("edc.agent.identity.key", "BusinessPartnerNumber");
                 }
             });
+    public static final String PLATO_NAME = "PLATO";
+    public static final String PLATO_BPN = PLATO_NAME + BPN_SUFFIX;
 
     @RegisterExtension
     protected static Participant plato = new Participant(
             ":edc-tests:runtime",
-            "PLATO",
+            PLATO_NAME,
+            PLATO_BPN,
             new HashMap<>() {
                 {
                     put("edc.connector.name", "plato");
-                    put("edc.ids.id", "urn:connector:plato");
+                    put("edc.participant.id", PLATO_BPN);
                     put("web.http.default.port", String.valueOf(PLATO_CONNECTOR_PORT));
                     put("web.http.default.path", PLATO_CONNECTOR_PATH);
                     put("web.http.management.port", String.valueOf(PLATO_MANAGEMENT_PORT));
                     put("web.http.management.path", PLATO_MANAGEMENT_PATH);
-                    put("web.http.ids.port", String.valueOf(PLATO_IDS_API_PORT));
-                    put("web.http.ids.path", IDS_PATH);
+                    put("web.http.protocol.port", String.valueOf(PLATO_DSP_API_PORT));
+                    put("web.http.protocol.path", DSP_PATH);
+                    put("edc.dsp.callback.address", PLATO_DSP_CALLBACK);
                     put("edc.api.auth.key", "testkey");
-                    put("ids.webhook.address", PLATO_IDS_API);
                     put("web.http.public.port", PLATO_PUBLIC_API_PORT);
                     put("web.http.public.path", "/api/public");
                     // embedded dataplane config
@@ -100,6 +107,7 @@ public class MultiRuntimeTest {
                     put("edc.dataplane.selector.httpplane.destinationtypes", "HttpProxy");
                     put("edc.dataplane.selector.httpplane.properties", "{\"publicApiUrl\":\"http://localhost:" + PLATO_PUBLIC_API_PORT + "/api/public\"}");
                     put("tractusx.businesspartnervalidation.log.agreement.validation", "true");
+                    put("edc.agent.identity.key", "BusinessPartnerNumber");
                 }
             });
 }